requirements.txt
requests
PyTelegramBotAPI
sqlalchemy
python-dotenv
